Transfer status using sender and client
